requirements.txt
flask
flask-cors
gunicorn
pandas
numpy
faiss-cpu
rank_bm25
google-generativeai
langchain
langchain-community
langchain-text-splitters
pypdf
unstructured
networkx
python-docx
python-dotenv